+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## v1.1.1 (08/12/2023)
+
+### Changed
+- Include username in get_consolidated_login_details
+- Upgrade requirements
+
+### fixed
+- Docstring download_entities
+
+
 ## v1.0.0 (28/11/2023)
 ### Added
 - Copy entities